EOS-18591: Basic implementation of remote execution module
- Add parsing logic
- Fix codacy issue
- SPec file addition
